Refactor seed node processes to new file.
